CibTools --> Cibtools
When renaming CibTools to Cibtools back then,
I mislooked some. Here I rename the rest.
HOWEVER. Maybe this change is unnecessary when using the
tumbleweed gems. (It has failed with the new upstream gems,
but now it's ok).
So if you want to keen this change you need to explain it,
like "gem ABC in a newer version requires it be Cibtools
rahter then CibTools"